src/main/java/com/trendyol/shoppingcard/controller/ProductController.java
package com.trendyol.shoppingcard.controller;

import com.trendyol.shoppingcard.common.ApiResponse;
import com.trendyol.shoppingcard.model.Product;
import com.trendyol.shoppingcard.service.ProductService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/product")
public class ProductController {

    private ProductService productService;

    @Autowired
    public ProductController(ProductService productService) {
        this.productService = productService;
    }

    @PostMapping("/save")
    public ApiResponse<Product> save(@RequestBody Product product) {
        return ApiResponse.response(productService.save(product));
    }

    @GetMapping("/")
    public ApiResponse<Product> get(String productId) {
        return ApiResponse.response(productService.get(productId));
    }

}
